+++ test/unit/routers/alpha-router/gas-models/v3-gas-model.test.ts
@@ -1,16 +1,13 @@
-import { Currency, CurrencyAmount, Ether, Token } from '@uniswap/sdk-core';
+import { Currency, CurrencyAmount, Ether } from '@uniswap/sdk-core';
 import { BigNumber } from 'ethers';
 import _ from 'lodash';
 import {
   DAI_MAINNET,
-  LiquidityCalculationPools,
   USDC_MAINNET,
   V3HeuristicGasModelFactory,
-  V3PoolProvider,
   V3Route,
   WRAPPED_NATIVE_CURRENCY,
 } from '../../../../../src';
-import { ProviderConfig } from '../../../../../src/providers/provider';
 import {
   BASE_SWAP_COST,
   COST_PER_HOP,
@@ -21,10 +18,6 @@ import {
   SINGLE_HOP_OVERHEAD,
 } from '../../../../../src/routers/alpha-router/gas-models/v3/gas-costs';
 import {
-  getHighestLiquidityV3NativePool,
-  getHighestLiquidityV3USDPool,
-} from '../../../../../src/util/gas-factory-helpers';
-import {
   DAI_USDT_LOW,
   USDC_USDT_MEDIUM,
   USDC_WETH_MEDIUM,
@@ -34,6 +27,7 @@ import {
   getMockedV2PoolProvider,
   getMockedV3PoolProvider,
 } from './test-util/mocked-dependencies';
+import { getPools } from './test-util/helpers';
 
 describe('v3 gas model tests', () => {
   const gasPriceWei = BigNumber.from(1000000000);
@@ -43,49 +37,6 @@ describe('v3 gas model tests', () => {
   const mockedV3PoolProvider = getMockedV3PoolProvider();
   const mockedV2PoolProvider = getMockedV2PoolProvider();
 
-  // helper function to get pools for building gas model
-  async function getPools(
-    amountToken: Token,
-    quoteToken: Token,
-    v3PoolProvider: V3PoolProvider,
-    providerConfig: ProviderConfig
-  ): Promise<LiquidityCalculationPools> {
-    const usdPoolPromise = getHighestLiquidityV3USDPool(
-      chainId,
-      v3PoolProvider,
-      providerConfig
-    );
-    const nativeCurrency = WRAPPED_NATIVE_CURRENCY[chainId];
-    const nativeQuoteTokenV3PoolPromise = !quoteToken.equals(nativeCurrency)
-      ? getHighestLiquidityV3NativePool(
-          quoteToken,
-          v3PoolProvider,
-          providerConfig
-        )
-      : Promise.resolve(null);
-    const nativeAmountTokenV3PoolPromise = !amountToken.equals(nativeCurrency)
-      ? getHighestLiquidityV3NativePool(
-          amountToken,
-          v3PoolProvider,
-          providerConfig
-        )
-      : Promise.resolve(null);
-
-    const [usdPool, nativeQuoteTokenV3Pool, nativeAmountTokenV3Pool] =
-      await Promise.all([
-        usdPoolPromise,
-        nativeQuoteTokenV3PoolPromise,
-        nativeAmountTokenV3PoolPromise,
-      ]);
-
-    const pools: LiquidityCalculationPools = {
-      usdPool: usdPool,
-      nativeQuoteTokenV3Pool: nativeQuoteTokenV3Pool,
-      nativeAmountTokenV3Pool: nativeAmountTokenV3Pool,
-    };
-    return pools;
-  }
-
   it('returns correct gas estimate for a v3 route | hops: 1 | ticks 1', async () => {
     const amountToken = USDC_MAINNET;
     const quoteToken = DAI_MAINNET;
@@ -94,7 +45,7 @@ describe('v3 gas model tests', () => {
       amountToken,
       quoteToken,
       mockedV3PoolProvider,
-      {}
+      {},
     );
 
     const v3GasModel = await v3GasModelFactory.buildGasModel({
